🐹📚 AtomicInt using custom SpinLock
Spin Lock concurrent pattern implementation applied to ensure atomicity of "int" operations.

As you can see, the result is unpredictable and also there are some side effects on chained operations between goroutines, but each of those operations are guaranteed to be atomic.

This lock type implemented using internally the CAS algorithm (compare and swap)

CAS algorithm is a well-known unlocked algorithm. Lock-free programming is to synchronize variables between threads without using locks, that is, to synchronize variables without thread blocking, so it is also called non-blocking Synchronization. The CAS algorithm involves three operands

a) Memory value V to read and write
b) Value A for comparison
c) New value B to be written

If and only if the value of V equals A, CAS updates the value of V with the new value B atomically, otherwise no operation will be performed (comparison and substitution is an atomic operation). In general, it is a spin operation, i.e. repeated retries.


Spinlock refers to when a thread acquires a lock, if the lock has been acquired by other threads, then the thread will wait for a loop, and then constantly determine whether it can be successfully acquired, until the lock is acquired, it will not exit the loop.

Threads that acquire locks are always active, but do not perform any valid tasks. Using such locks results in busy-waiting.

It is a lock mechanism proposed to protect shared resources. In fact, spin locks are similar to mutexes in that they are used to solve the mutex of a resource. Whether mutex or spin lock, at any time, only one holder can get the lock at most, that is to say, only one execution unit can get the lock at any time. But they are slightly different in scheduling mechanism. For mutexes, if resources have been occupied, resource applicants can only sleep. But the spin lock does not cause the caller to sleep. If the spin lock has been maintained by other executing units, the caller keeps looping there to see if the holder of the spin lock has released the lock. That is why the term “spin” is named.


Then...When to SpinLock? When to Mutex?
Differences?

* SpinLock: As you can see, SpinLock idea comes from Loop-CAS, which is lock-free, but kinda busy checking.
* Mutex: When Lock() invocation happens, however can’t get the lock, the thread/goroutine will be added to wait list of scheduler and thread/goroutine switch happens at this time.

So SpinLock thinks that I should very probably get the lock in my scheduling CPU slot, and in a aggressive way.

What about Mutex scheduling? In Golang implementation, it’s:

1. If Lock() happens and can’t get the lock, will be added to wait list,
2. For new young goroutine who already get the lock, will just let them run, since they already get the lock,
3. If the old goroutine has been wait for very long time, will give the lock to him and let him run, so it won’t wait for too long. By default, starvationThresholdNs is 1e6 nanoseconds (1ms). Normally it’s FIFO order, but if starvationThresholdNs meet, will enter starvation mode.

In a summary:

1. Under multi-core CPU, if you think that you will quickly get the lock, then please use SpinLock, so that it doesn’t need context switch to get the lock and execute the code.
  1a. the time to get the lock is short (short task or low concurrency)
  1b. the critical code block running time cost won’t be too much, i.e., can quickly do the shared variables computation and release the lock
  1c. the concurrency is not too much, or it will needs lots of context switches to release the lock

2. Under single-core CPU, never use SpinLock, since it will try to use up the CPU slot, until context switch happens and then back to get the lock. Context Switch always happens in 1 Core CPU. But the above SpinLock code implementation is good, since it always yield the CPU to the other goroutines by using runtime.Gosched(). So the SpinLock implementation here is a bit like a Mutex but lock-free.

3. Under multi-core CPU, if you think your critical code block can’t complete in the assigned CPU slot, please try to use Mutex, so that it won’t waste up much time on getting the lock.
  3a. the time to get the lock is long (long task or high concurrency)
  3b. the critical code block running time cost is little large, i.e, keeps computation on top of shared variables for a long time. So it’s important not to put I/O into critical code block.

To sum up:

1. Single-core: not use SpinLock, or at least yield the processor every time failed CAS.
2. Multi-core, short task or low concurrency: use SpinLock.
3. Multi-core, long task or high concurrency: use Mutex.
4. Don't use I/O in critical code block, this should be a common knowledge.

If high concurrency with heavy I/O, then better to use goroutine + channel solution.