🐹👑 Office Computers (Math Operators 2)
Office Computers. Math Operators +10 XP

In the office, 2 monitors are connected to each computer.
The first line of the given code takes the number of computers as input.

Task
Complete the code to calculate and output the number of monitors to the console.

Sample Input
10

Sample Output
20

Hint
Since each computer has 2 monitors, simply multiply the count of the computers by 2.

Use the multiplication operator (*).