🐹👑 Find the adults (Comparison Math Operators)
Find the adults. Comparison Math Operators +10 XP

check someone’s age...not just at the bar!

You are given a program that takes the age of the user as input.
Complete the code to check if the user is an adult, and output to the console the corresponding boolean value.

Sample Input
20

Sample Output
true

Hint:
If the user is 18 or older, they’re considered an adult.
fmt.Println(20>18) outputs true.