🐹👑 Vacation Month
Vacation Month
The if Statement +10 XP

You are planning a vacation in August.
You are given a program that takes the month as input.

Task
Complete the code to output "vacation", if the given month is August. Don't output anything otherwise.

Sample Input
August

Sample Output
vacation

Hint:
Handle the required condition using an if statement.
Use fmt.Println() for the output.