🐹👑 Dark Theme Menu
Dark Theme
Switch statements +10 XP

The user can choose the color theme for the browser:
1. Light
2. Dark
3. Nocturne
4. Terminal
5. Indigo

You are given a program that takes the number as input. Complete the program so that it will output to the console the theme according to input number.

Sample Input
2

Sample Output
Dark

Hint:
Don't forget to use break for each case statement.