🐹👑 Repeat and Learn Code! Code! Code!
Repeat and Learn Code! Code! Code!
The for loop 2. +10 XP

"Repetition is the mother of learning, the father of action, which makes it the architect of accomplishment." - Zig Ziglar.
Inspired by these words, let's create a little program that will output an expression which is given as input, 3 times.

Task
Complete the code to output the given expression 3 times.

Sample Input
Learning is fun!

Sample Output
Learning is fun!
Learning is fun!
Learning is fun!

Hint:
Use the for loop to run the required part of the code as many times as needed.