🐹👑 Skipping 13 reflect
Skipping 13
Break and Continue. +10 XP

Many tall buildings, including hotels, skip the number 13 when numbering floors -- often going from floor 12 to floor 14. It is thought that the number 13 is unlucky.
Write a program that will number 15 rooms starting from 1, skipping the number 13. Output to the console each room number in separate line.

Hint:
Be attentive - considering the missing number, the number of last room should be greater than the count of rooms by 1.