🐹👑 Average Of Three
Average Of Three
The return Statement. +10 XP

You are given a program that takes 3 numbers as input.
Complete the given function to calculate the average of those 3 numbers, assign it to the given variable, and output it.

Sample Input
3
6
9

Sample Output
6

Hint:
Use return statement to return the calculated value and use it through the program.