🐹👑 Currency Converter - OOP big.Float
Currency Converter. +50 XP

You are making a currency converter app.
Create a function called convert, which takes two parameters: the amount to convert, and the rate, and returns the resulting amount.
The code to take the parameters as input and call the function is already present in the Playground.
Create the function to make the code work.

Sample Input:
100
1.1

Sample Output:
110

Hint:
Converting 100 at the rate of 1.1 is equal to 100*1.1 = 110.