🐹👑 Cuboid Volume OOP with panicCatcher
Cuboid Volume. +10 XP
Introducing Objects

The given structure represents a cuboid (e.g. a rectangular prism) that contains the properties of length, width, and height.
Complete the program to calculate and output the volume of given cuboid to the console.

Hint
To calculate the volume of cuboid use length*width*height formula.